Chinese Console version 0.2.3.
See https://sourceforge.net/projects/zhcon/ .

Significant patching was required, to fix bugs caught by the clang++
compiler on FreeBSD, to cleanup problems where certain control sequences
simply cause the entire program to abend, and to fix the unfortunate
default choice of a screen resolution that the FreeBSD framebuffer
drivers cannot in fact mmap() properly.

Running from the package directory directly simply requires either symbolic
links in relevant /command and manual directories or additions to the
relevant command and manual search paths:
	PATH=${PATH:-${PATH}:}/package/sysutils/zhcon-0.2.3/command
	MANPATH=${MANPATH:-${MANPATH}:}/package/sysutils/zhcon-0.2.3/manual

The configuration files have not been patched to fix the default encoding
choice of GB2312, which is not nearly as popular now as in 2015 when this
version was released.